Move update-related tests from snapstate_test.go to snapstate_update_test.go.
